# cmd/wtd/openapi.json
{
 "components": {
  "responses": {
   "BadRequest": {
    "content": {
     "application/json": {
      "examples": {
       "badName": {
        "value": {
         "error": {
          "code": "invalid_name",
          "message": "name must be 1-64 characters of [A-Za-z0-9._-], must not start with '.', and must not contain '..'"
         }
        }
       }
      },
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    },
    "description": "Validation failure; error.code disambiguates (invalid_name, invalid_path, unknown_project, project_path_missing, path_and_project, bad_request)."
   },
   "Internal": {
    "content": {
     "application/json": {
      "examples": {
       "oops": {
        "value": {
         "error": {
          "code": "internal",
          "message": "unexpected error; see journal"
         }
        }
       }
      },
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    },
    "description": "Unexpected server fault."
   },
...
  },
  "schemas": {
...
   "Session": {
    "description": "One session. How each field is derived, and how each can fail: `/docs/session-lifecycle.md` \u00a72.\n\n**Nulls are per-response, not sticky.** Every field is re-derived from the server's filesystem and process table on each request \u2014 nothing is cached \u2014 so a session reporting `cwd: null` or `pid: null` now may carry values on the next request, and the reverse. A null is never a permanent property of the session. It is equally not a promise that retrying helps: the usual cause is a process this server cannot inspect, and that persists for the session's lifetime. So do not poll specially and do not show a spinner waiting for a value to arrive \u2014 render the absence, and let your next ordinary refresh fill it in if it can.\n",
    "properties": {
     "agentStatus": {
      "description": "The last agent status this server observed on the session's own output, as defined by the in-band convention in `/docs/ws-protocol.md` \u00a76a. Present only when `features` includes `session-status`.\n\n**Null means unknown, and unknown is not idle.** Null says this server has never observed this session's output \u2014 no client has opened it through this server \u2014 so it is a statement about the server, not about the session. A session used only over SSH stays null forever, and every session is null after a restart until someone opens it. `clear` is the opposite kind of fact: the session itself reporting that it has nothing to report. Render the two differently, or a row nobody has looked at will claim nothing is happening.\n\n**A status does not expire, so it can outlive what reported it.** \u00a76a's limitation applies unchanged here: a state stands until something reports a different one, so an emitter killed rather than exited leaves `running` pinned. Do not add a timeout to work around this \u2014 \u00a76a forbids clients inventing one, and the reasoning is there.\n\nIgnore an unrecognized value rather than treating it as `clear`, so a server that learns a new state does not blank your indicator.\n",
      "enum": [
       "running",
       "waiting",
       "attention",
       "clear",
       null
      ],
      "type": [
       "string",
       "null"
      ]
     },
     "attached": {
      "description": "True while someone is looking at this session \u2014 through this server or over SSH, so `true` does not imply a client of *this* server. That sentence is the contract. How it is worked out is not, and has already changed once, so clients MUST NOT derive it themselves and in particular MUST NOT read the socket's file mode. The derivation, and why it changed: `/docs/session-lifecycle.md` \u00a72.\n",
      "type": "boolean"
     },
     "attachedCount": {
      "description": "How many viewers `attached` is counting \u2014 the same population, so this is `0` whenever `attached` is false, and at least `1` whenever it is true and this is not null.\nNull means someone is attached and the server could not count them, which happens when the only evidence is dtach's own signalling and the viewer's process is not one this server can inspect. Treat it as \"one or more\", never as zero. Null is possible only while `attached` is true.\nThe counted case is best-effort rather than a census. Viewers connected through this server are exact. Viewers attached by other means are identified from the server's process table, which can miss one it is not allowed to inspect, and can briefly count a session's own supervising process while that session is shutting down. So use this to tell \"just me\" from \"someone else as well\"; do not build on an exact value, and do not treat a change in it as an event.\nThere is no way to ask which of these viewers is *you* \u2014 this server issues no client or connection identity, so nothing in this API could name one, and that is not planned. The usable substitute is arithmetic: a client that is itself attached to the session knows one of the count is its own, so `1` means it is alone and `2` or more means someone else is watching too.\n",
      "minimum": 0,
      "type": [
       "integer",
       "null"
      ]
     },
     "createdAt": {
      "description": "RFC 3339 UTC. Effectively when the session was created, but best-effort rather than forensic \u2014 treat it as ordering information, not evidence.\n",
      "format": "date-time",
      "type": "string"
     },
     "cwd": {
      "description": "Current working directory of the session's shell, read live \u2014 it moves when the user `cd`s, so this is not the directory the session was created in. Null when it cannot be resolved; the session is still valid and still listed.\n",
      "type": [
       "string",
       "null"
      ]
     },
     "name": {
      "description": "Deliberately unconstrained here: a deep link creates names `POST` would reject, such as ones containing spaces, and this API lists every session that exists. Do not validate a listed name against the create-side pattern.\nAlways valid UTF-8, and never contains `/` \u2014 a name is a filename component on the server, so one cannot occur. That matters because this string is the byte-exact key for `GET` and `DELETE` on `/api/v1/sessions/{name}`: round-trip it unchanged, URL-encoding as the transport requires, and it will always match.\n",
      "type": "string"
     },
     "pid": {
      "description": "PID of the session's shell \u2014 the same process whose `cwd` is reported above. Null when it cannot be resolved; the session is still valid and still listed.\n",
      "type": [
       "integer",
       "null"
      ]
     }
    },
    "required": [
     "name",
     "attached",
     "attachedCount",
     "cwd",
     "pid",
     "createdAt",
     "agentStatus"
    ],
    "type": "object"
   },
...
  }
 },
...
 "info": {
...
  "license": {
   "identifier": "MIT",
   "name": "MIT"
  },
  "summary": "Browser terminal daemon: ttyd-wire-compatible WebSocket terminal plus a JSON API over dtach sessions.\n",
  "title": "wtd HTTP API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/sessions": {
   "get": {
    "description": "Every session this server can see, sorted ascending by the byte value of `name`. Sessions\ncan be created by deep link as well as by `POST`, and every surface that lists them must\nagree about what exists \u2014 which drives both rules below.\n\n1. **Looser names than this API would create are still listed.** A session made by\n   deep-linking `/ws?arg=my%20proj` is named `my proj`; it appears here even though\n   `POST` would reject that name. Do not assume listed names satisfy the create-side\n   pattern. Names beginning with `.` are listed too, even though `POST` refuses them.\n2. **Dead sessions are not listed, and this request removes them.** A socket\n   left behind by a session that is gone would otherwise show as a session\n   that cannot be attached, so this endpoint verifies each one and unlinks the\n   ones it can prove are dead. That probing is also why this read is not instant:\n   budget a client timeout of a few seconds rather than a few hundred milliseconds, and\n   expect the cost to grow with the number of sessions. Yes, that means a `GET` has a\n   side effect; it is\n   the only read every client already makes, which is why it is the one that\n   converges what exists with what is listed. Why, and the guard against racing a\n   session being recreated under the same name: `/docs/session-lifecycle.md` \u00a75.\n",
    "operationId": "listSessions",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "examples": {
         "two": {
          "value": [
           {
            "attached": false,
            "attachedCount": 0,
            "createdAt": "2026-07-22T21:03:00Z",
            "cwd": "/home/sam/src/infra",
            "name": "ops",
            "pid": 51234
           },
           {
            "attached": true,
            "attachedCount": 2,
            "createdAt": "2026-07-24T14:09:00Z",
            "cwd": "/home/sam/src/riggs",
            "name": "riggs",
            "pid": 60211
           }
          ]
         }
        },
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Session"
         },
         "type": "array"
        }
       }
      },
      "description": "All live sessions (possibly empty array).",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-store",
         "type": "string"
        }
       }
      }
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "List dtach sessions"
   },
   "post": {
    "description": "Creates a dtach session **without attaching**, for naming a working directory or\nproject up front. It is **not** a precondition for deep-linking: `/ws?arg=<name>`\ncreates the session if it does not exist, so a client can open one directly and never\ncall this.\n\nA session created here is **indistinguishable** from one created by deep link \u2014 same\ninvocation, same environment, built from the same code. That parity is mandatory, so\nanything you can do with a deep-linked session you can do with this one, and nothing\nabout it will surprise a user who switches between the two.\nThe invocation and the reasons each part of it matters: `/docs/session-lifecycle.md` \u00a76.\n\nWorking directory: `project` (a shortcut name from `GET /api/v1/projects`) or\n`path` (absolute, must exist), else the service user's `$HOME`. The two are\nmutually exclusive. Where a deep-linked connection silently falls back to `$HOME`\nfor a directory that is missing, this API returns an error instead \u2014 a program\nshould be told, not guessed for.\n\nName rules are **stricter than the deep-link path's**, deliberately: no leading `.`\nand no `..` anywhere, the latter because such a session could never be reached by\ndeep link. See the `name` property for the length ceiling, which is\ndeployment-dependent.\n\nIf the name belongs to a session that is already gone, its leftover socket is\nremoved and creation proceeds \u2014 so retrying a create for a session that has\ndied does the obvious thing rather than returning a conflict.\n",
    "operationId": "createSession",
    "requestBody": {
     "content": {
      "application/json": {
       "examples": {
        "bare": {
         "value": {
          "name": "scratch"
         }
        },
        "byPath": {
         "value": {
          "name": "riggs",
          "path": "/home/sam/src/riggs"
         }
        },
        "byProject": {
         "value": {
          "name": "ops",
          "project": "ops"
         }
        }
       },
       "schema": {
        "$ref": "#/components/schemas/SessionCreateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "examples": {
         "created": {
          "value": {
           "attached": false,
           "attachedCount": 0,
           "createdAt": "2026-07-24T21:00:00Z",
           "cwd": "/home/sam/src/riggs",
           "name": "riggs",
           "pid": 61402
          }
         }
        },
        "schema": {
         "$ref": "#/components/schemas/Session"
        }
       }
      },
      "description": "Session created, detached.",
      "headers": {
       "Location": {
        "description": "Canonical resource path for the new session.",
        "schema": {
         "examples": [
          "/api/v1/sessions/riggs"
         ],
         "type": "string"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/BadRequest"
     },
     "403": {
      "$ref": "#/components/responses/OriginForbidden"
     },
     "409": {
      "content": {
       "application/json": {
        "examples": {
         "exists": {
          "value": {
           "error": {
            "code": "already_exists",
            "message": "session 'riggs' already exists"
           }
          }
         }
        },
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "A session with this name already exists. Two concurrent creates of the same name resolve to one winner and one 409. A session whose death cannot be established counts as existing, and is reported rather than silently replaced."
     },
     "413": {
      "$ref": "#/components/responses/PayloadTooLarge"
     },
     "415": {
      "$ref": "#/components/responses/UnsupportedMediaType"
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "Create a detached session"
   }
  },
...
  "/docs/{file}": {
   "get": {
    "description": "The documents this spec references, served so those references resolve. OpenAPI cannot express a WebSocket protocol, so `/docs/ws-protocol.md` is where the terminal's wire contract is stated in full; `/docs/session-lifecycle.md` carries the session model and `/docs/compatibility.md` the ttyd behaviour matrix. Markdown, intended to be read. Detect with the `docs-endpoint` feature.\n",
    "operationId": "getDoc",
    "parameters": [
     {
      "description": "One of the published document names.",
      "in": "path",
      "name": "file",
      "required": true,
      "schema": {
       "enum": [
        "ws-protocol.md",
        "session-lifecycle.md",
        "compatibility.md"
       ],
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "text/markdown": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": "The document.",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-cache",
         "type": "string"
        }
       }
      }
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "No document by that name."
     }
    },
    "summary": "A specification document this server publishes"
   }
  },
...
  "/token": {
   "get": {
    "description": "Exists solely because clients of ttyd expect it: they GET it on every connect,\nignore failures, and send the result as the handshake `AuthToken`. wtd has no\nauth, so the token is always the empty string \u2014 the same as ttyd run without\n`-c`. Serving it keeps connect logs clean and third-party ttyd-compatible\nclients working.\n",
    "operationId": "getToken",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "examples": {
         "empty": {
          "value": {
           "token": ""
          }
         }
        },
        "schema": {
         "properties": {
          "token": {
           "const": "",
           "type": "string"
          }
         },
         "required": [
          "token"
         ],
         "type": "object"
        }
       }
      },
      "description": "The (empty) token.",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-store",
         "type": "string"
        }
       }
      }
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "ttyd-compatible auth token endpoint (always empty)"
   }
  },
  "/ws": {
   "get": {
    "description": "Not an HTTP resource in any REST sense \u2014 this path upgrades to the ttyd\n`tty`-subprotocol WebSocket. OpenAPI cannot express a WebSocket protocol, so what\nfollows is the working contract in prose. It is enough to build a client; the full\nnormative version, including every failure mode, is `/docs/ws-protocol.md`.\n\n## Connecting\n\nOpen `ws://<host>:<port>/ws`, offering subprotocol `tty`. The server echoes it, which\nbrowsers require. Offering **none** also works and the connection proceeds with no\nsubprotocol selected; offering others **without** `tty` is a `400` before the upgrade,\nso check the status rather than waiting for a close code. `GET /token` first if you want ttyd parity \u2014 its value goes in the handshake \u2014\nbut the token here is always the empty string.\n\n## The handshake\n\nThe **first** message after the upgrade is the handshake, and it is the one message with\n**no opcode prefix**. Send `TerminalHandshake` as **either a text or a binary** frame;\nboth are accepted, which matters because ttyd's own web client sends binary and other\nclients send text.\n\nSend real `columns`/`rows`, as **numbers**. Missing or out-of-range values (outside\n**1..65535**) default to **80x25 silently**, and the size is applied before the start\ncommand runs \u2014 so a narrow client that omits them gets a session hard-wrapped for 80\ncolumns with no error anywhere. A dimension that is not a number at all (`\"80\"`, `80.5`,\n`true`) is not defaulted: it makes the payload undecodable and closes **1002**.\n\nUnknown keys are ignored. Not-JSON closes **1002**; no handshake within **10 s** closes\n**1008**; over **8 KiB** closes **1009**.\n\n## Framing, after the handshake\n\nEvery subsequent message is `[1 ASCII opcode byte][payload]`, applied to the reassembled\nmessage rather than to transport fragments.\n\n| Client \u2192 server | Byte | Payload |\n|---|---|---|\n| `INPUT` | `'0'` | Raw bytes, written to the terminal verbatim. No escaping, no encoding. |\n| `RESIZE_TERMINAL` | `'1'` | `TerminalResize` as JSON. |\n| `PAUSE` / `RESUME` | `'2'` / `'3'` | Accepted for ttyd parity and **no-ops here**; see the `flow-control` feature. |\n\n| Server \u2192 client | Byte | Payload |\n|---|---|---|\n| `OUTPUT` | `'0'` | Raw terminal bytes. |\n| `SET_WINDOW_TITLE` | `'1'` | UTF-8 string; display verbatim, nothing parses it. |\n| `SET_PREFERENCES` | `'2'` | JSON, always `{}` here. |\n\nClient messages may be text or binary. **Server messages are always binary.**\n\n## Three rules that silently corrupt rendering\n\n1. **The first two server frames are the title and the preferences**, in that order,\n   before any output. Feed them to a terminal emulator and it prints them on screen.\n2. **`OUTPUT` is not UTF-8-safe per frame.** A multi-byte character can split across two\n   frames. Decode the concatenated stream, never one frame at a time, or non-ASCII output\n   corrupts intermittently at frame boundaries.\n3. **Ignore unknown opcodes; do not close on them.** New opcodes may appear, and a client\n   that treats byte 0 as an error never gets past its first frame. The server ignores\n   unknown client opcodes for the same reason.\n\n## Two kinds of connection, chosen by `?arg=` alone\n\nNothing else on the wire differs.\n\n| | argless `/ws` | named `/ws?arg=<name>` |\n|---|---|---|\n| Process | fresh per connection, private shell | one shared session, kept alive between connections |\n| On attach | a new shell prompt | recent output replayed, then live |\n| A second client | gets its own shell | joins the same terminal, **input interleaves** |\n| You disconnect | the process is terminated | **the session keeps running** |\n\n`?arg=<name>` **creates the session if it does not exist** and recovers one whose previous\nprocess is gone \u2014 so a saved deep link works before its session has ever existed, and\nafter a reboot. You do not need to `POST` first.\n\nAn **empty** `?arg=` counts as argless: no replay, no sharing, and the process dies with\nthe connection. Build the URL without the parameter rather than with an empty one.\n\n**An argless connection is not how you browse sessions.** It is a plain shell in the\nservice user's home directory, with no session behind it \u2014 nothing to reattach to later.\nUse `GET /api/v1/sessions` to list, and `?arg=<name>` to open one.\n\nA name that cannot be used still opens a working terminal \u2014 never an error, never a close.\nSee the `arg` parameter for exactly what that terminal is.\n\n**Ctrl-`\\` is shared fate on a named session.** It detaches the session, which closes\n**every** client attached to it, not only the one that sent it. Indistinguishable from a\nprivate detach when you are the only client.\n\n## Resize, and the redraw you should not send\n\nEvery `RESIZE_TERMINAL` frame is applied immediately, one per frame, never coalesced \u2014\nso debouncing an interactive resize is your job. Dimensions outside 1..65535 are ignored\nrather than fatal. With several clients on one session the size is last-writer-wins:\njoining a session resizes it for everyone already there.\n\n**You do not need to send a resize to get output**, and you should not send a\nsize-jiggle to force a redraw. This server performs that once per session itself. A\nclient-side jiggle still works but repaints the screen for **every** client on a shared\nsession. This is the one rule that inverts against a plain ttyd server, which does no\nredraw of its own and where a client-side jiggle is the only thing preventing a blank\nscreen on attach \u2014 so gate it on the `scrollback-replay` feature rather than assuming.\n\n## Replay, when `scrollback-replay` is advertised\n\nRecent output is replayed after the title and preferences frames, before live output.\nIt arrives as **ordinary `OUTPUT` frames with no marker of any kind** \u2014 deliberately, so\nthat a client which knows nothing about replay still works. Consequences:\n\n- **You cannot detect where replay ends.** Do not try.\n- **It overlaps whatever you already have** if you reconnect into a terminal emulator you\n  kept alive. Clearing before reconnecting is your decision; the server injects nothing.\n- **It is wrapped at the width of the session that produced it**, not yours. A narrow\n  client joining a wide session sees history wrapped for the wide one. Full-screen\n  programs repaint; a shell's scrollback does not.\n- **Do not depend on it to render a usable screen.** An operator can disable it while the\n  feature stays advertised.\n\nReplay covers what happened *before* you attached; your own scrollback covers what\narrived *while* you were connected. Different mechanisms, both needed.\n\n## Liveness\n\n**You do not have to ping** \u2014 browsers cannot, so the server never treats silence from a\nclient as death. It pings every 30 s and closes after **three consecutive pings go\nunanswered**, so a live connection has 90 s of tolerance and a single lost pong costs\nnothing. Transport-level auto-pong satisfies this for free: browsers and URLSession both\nanswer without any code of yours, which is why an idle terminal survives indefinitely.\n\nWhat is consulted is the **pong**, not traffic in general \u2014 sending input does not by\nitself keep a connection alive if your transport is not answering pings. No real client\nis in that position, and a hand-rolled one that suppresses pongs is the case to know\nabout. There is no idle timeout below 60 s. The server answers pings with pongs and\nignores unsolicited pongs.\n\n## Limits\n\nAny message after the handshake over **1 MiB** closes **1009**. That is the ceiling on a\nsingle paste. It is unrelated to the 16 KiB cap on HTTP request bodies documented above.\n\nWhat else it takes down depends on the connection shape. On an **argless** connection the\nprocess is yours alone and dies with you. On a **named** one only your client is dropped \u2014\nthe session keeps running for every other client, so reconnecting lands you back in it with\nreplay. One client's oversized paste cannot end a session someone else is using.\n\n## Close codes, and what to do about each\n\n| Code | Means | Do |\n|---|---|---|\n| `1000` | The session's command exited. On a named session this is the shared command, so every client gets it. | Final. Reconnect only if the user asks. |\n| `1001` | Server shutting down, or a warm session released to make room. Sent on **named** connections; an argless one is dropped without a close frame on shutdown. | Reconnect with backoff. Sessions survive a restart. |\n| `1013` | You fell more than 4 MiB behind on output and were dropped so the session could keep going. | Reconnect immediately; replay restores context. |\n| `1002` | The first message was not a valid handshake. | Client bug. Do not retry. |\n| `1008` | No handshake within 10 s. | Client bug. Do not retry. |\n| `1009` | A message exceeded 8 KiB (handshake) or 1 MiB (after). | Client bug. The same payload fails identically. |\n| `1011` | The terminal could not be started. The title and preferences frames arrive first, then an `OUTPUT` frame with a one-line error, so the reason is visible in the terminal rather than only in a code. | Retry with backoff \u2014 but display the message. The commonest cause is a misconfigured start command, which fails the same way every time. |\n\n`TERM=xterm-256color` is set, so xterm-family emulation is the target.\n",
    "externalDocs": {
     "description": "The full normative wire contract, including everything this description summarises.",
     "url": "/docs/ws-protocol.md"
    },
    "operationId": "terminalWebSocket",
    "parameters": [
     {
      "description": "Session name. Present and non-empty selects a named connection, creating the session if it does not exist; absent or empty selects a private one. Repeats are passed through in order, though only the first is used today. URL-decoded before use.\nA name this server cannot build a session from does not fail the connection: it opens a plain shell instead, preceded by one `OUTPUT` line beginning `wtd:` that says why. **No value of `arg` closes the connection.** Two kinds of unusable name behave differently afterwards, and the difference is observable if a second client deep-links the same name:\n*Unusable as a session, still a name.* One containing `/` or `..`, and one whose socket path would exceed the operating system's limit. No session is created, but the connection remains a *named* one for sharing and replay \u2014 so two clients on the same unusable name land on **one shared fallback shell**, and both see the explanation.\n*Not carriable at all, so discarded.* A value containing a NUL byte (`%00`) or over **4096 bytes**. These cannot be carried as a process argument, so the server drops them before choosing a session. Separately, only the first **16 usable** values are kept and any beyond that are ignored \u2014 counted after the unusable ones are removed, so sending one NUL does not cost a usable value its place. A connection whose only `arg` was dropped is therefore **argless**: private shell, no sharing, no replay, no explanation, and the process ends with the connection. Two clients doing it get one shell each.\nBoth are graceful; only the second loses the name, and only the first tells the user anything. Send a valid name if you want the sharing semantics.\nThe length ceiling is the one `POST /api/v1/sessions` enforces, so `maxSessionNameLength` from `GET /api/v1/meta` is a safe bound for both \u2014 validate against it before deep-linking, or your user gets a bare shell where they expected their session. Count **bytes of the UTF-8 encoding**, not characters: both the server and the operating system limit measure bytes, so a name that is short in characters and long in bytes is refused on the byte count. The reverse does not hold \u2014 this path accepts names `POST` would refuse, deliberately, so that a session created by any means can be reached by deep link.\nIgnored repeats are dropped rather than refused, so a client cannot tell from the wire that it sent too many. Nothing legitimate sends more than one.\n**Use the name exactly as `GET /api/v1/sessions` returned it.** That list is byte-exact, and this path accepts names `POST` would refuse \u2014 an interior space is legitimate and common (`my project`), and **leading or trailing whitespace is possible too**, because nothing between the URL and the socket name trims it. Trimming or normalising a name before putting it in `?arg=` therefore attaches to, or creates, a *different* session than the one the user picked. Round-trip it verbatim.\n",
      "in": "query",
      "name": "arg",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "101": {
      "description": "Switching Protocols; subprotocol `tty` echoed when offered."
     },
     "400": {
      "description": "Not a valid upgrade, or subprotocols offered without `tty`."
     },
     "403": {
      "description": "Cross-origin upgrade rejected (`/docs/ws-protocol.md` section 12)."
     }
    },
    "summary": "Terminal WebSocket (ttyd wire protocol)"
   }
  }
 },
 "servers": [
  {
   "description": "wtd's single listener, and the only server that serves this document. host is the resolved bind address \u2014 a tailnet IP, 127.0.0.1, or an interface address, never a wildcard.\n**Use port 7681 (`WT_PORT`).** One box runs one listener on one port, and it serves every route in this document. A server that answers `/`, `/token` and `/ws` but returns 404 for `/openapi.json`, the `/api/v1/*` routes and the picker is an older deployment that has not been upgraded; `GET /api/v1/meta` succeeding is the cheapest proof you are talking to wtd, and its `features[]` is how to test for anything newer than this revision.\n",
   "url": "http://{host}:{port}",
   "variables": {
    "host": {
     "default": "127.0.0.1"
    },
    "port": {
     "default": "7681"
    }
   }
  }
 ]
}